Alright, here is my plan to fix the map opacity endpoint:
- I'll start by creating a new public API endpoint for map opacity.
- Then, I'll refactor the `get_map_opacity_value` function.
- Next, I'll update the frontend to use the new endpoint.
- After that, I'll update the admin API to use the refactored function.
- Finally, I'll register the new public API blueprint.
